+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2018-04-09
+### Added
+- More options on notes to edit, delete and print.
+
+### Fixed
+- Both common parenthesis and full width parenthesis work now after brackets.
+- Inserting "<" makes your text dissapear.
+
 ## [1.0.1] - 2018-04-02
 ### Added
 - Added missing theme color.